Revert "Fix SSL setup for haproxy 1.4 release"
We have haproxy 1.5.x now, so we can use persistence for SSL sessions.
This matters as if we want to allow people to not use "source" as
balance algorithm, then we need to make sure that sessions keep going to
the same backend to avoid breakages.

This reverts commit 32323b0a4b17d64489735d307b9fe65b8bb4ed31.

(cherry picked from commit 8cf8ddbd9e7813ee58b593dddb719aee0aaba91c)